hw5 done

+++ src/services/contacts.js
@@ -10,11 +10,12 @@ export const getAllContacts = async ({
   sortOrder = SORT_ORDER.ASC,
   sortBy = '_id',
   filter = {},
+  userId,
 }) => {
   const limit = perPage;
   const skip = (page - 1) * perPage;
 
-  const contactsQuery = ContactCollection.find();
+  const contactsQuery = ContactCollection.find({ userId });
 
   if (filter.contactType) {
     contactsQuery.where('contactType').equals(filter.contactType);
@@ -35,8 +36,8 @@ export const getAllContacts = async ({
   };
 };
 
-export const getContactsById = async (contactId) => {
-  const contact = await ContactCollection.findById(contactId);
+export const getContactsById = async (contactId, userId) => {
+  const contact = await ContactCollection.findOne({ _id: contactId, userId });
   return contact;
 };
 
@@ -44,8 +45,8 @@ export const createContact = async (payload) => {
   const contact = await ContactCollection.create(payload);
   return contact;
 };
-export const patchContact = async (contactId, payload, options = {}) => {
-  const rawResult = await ContactCollection.findOneAndUpdate({ _id: contactId }, payload, {
+export const patchContact = async (contactId, payload, userId, options = {}) => {
+  const rawResult = await ContactCollection.findOneAndUpdate({ _id: contactId, userId }, payload, {
     new: true,
     includeResultMetadata: true,
     ...options,
@@ -59,7 +60,7 @@ export const patchContact = async (contactId, payload, options = {}) => {
   };
 };
 
-export const deleteContact = async (contactId) => {
-  const contact = await ContactCollection.findOneAndDelete({ _id: contactId });
+export const deleteContact = async (contactId, userId) => {
+  const contact = await ContactCollection.findOneAndDelete({ _id: contactId, userId });
   return contact;
 };
